requirements.txt
# Dépendances système
python-dotenv==1.1.1
setuptools==80.9.0

# Librairies de base
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
requests==2.32.4
click==8.2.1
colorama==0.4.6
tqdm==4.67.1
filelock==3.18.0
platformdirs==4.3.8

# WandB pour le suivi des entraînements
wandb==0.21.0

# Machine Learning
numpy==2.3.2

# Modélisation & configuration
pydantic==2.11.7
pydantic_core==2.33.2
typing_extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# Manipulation de fichiers
fsspec==2025.7.0
PyYAML==6.0.2

# Logging / monitoring
sentry-sdk==2.33.2
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

# Tokenization et NLP
sentencepiece==0.2.0

# Math / graph / symbolic
sympy==1.14.0
mpmath==1.3.0
networkx==3.5

# Templates (si utilisé)
Jinja2==3.1.6
MarkupSafe==3.0.2
